Script that computes the discretized trajectory
To define the clusters it uses the last clustering object of
the adaptive run.

It assigns the snapshot to the cluster with closest RMSD. Note
that this might differ with the sort in the adaptive sampling
project, where we assign it to the first one that is close enough
(according to the threshold criterion)

It is optimised by computing the center of mass of the ligand:
if the COM distance is larger than the smallest rmsd so far, it
does not compute the rmsd.

It could be further optimised by adding a cell list to keep track
of the cluster's COM position.