Clustering optimisation using centroid distance before RMSD calc
It will compute the centroid distance between every snapshot and
the cluster center. If the distance is less or equal to the target
threshold RMSD, it will compute it to further check if the snapshot
belongs to the cluster or not.

It could be further applied an optimisation with a cell list, in order
to only get the cluster centers in neighboring cells.