Fixed bug of symmetry corrected RMSD
Moved the RMSD calculation into a class called RMSD calculator. This class is
in a module with the same name in atomset. Now the symmetries are a list of
dictionaries with a dictionary per group symmetry. The dictonary must only
contain a pair of key values per symmetric pair. Therefore the function
generateReciprocalDict is elimanated and the assertSimetryDict has been changed
in adaptiveSampling module. The tests and clustering have been adapted to the
new RMSDCalculator class.