Added brand new plotRelativeEntropy and imporved computeRelativeEntropy's performance.
plotRelativeEntropy takes the same controlFile as computeRelativeEntropy
and replots the relative entropy, being able to use the same scale as another
relative entropy plot. It nees to be improved, as the read npy file needs
to be changed in the code.

computeRelativeEntropy is improved, since the assignation is only done once for
each dataset, and is stored as a *.npy file. Then, it just bootstraps the
assigned data and computes the relative entropy, which is much faster if one only
wants to change the range if trajectories and lengths for example.
Besides, the number of trajectories and their length is now read form the control file