Add conformation network, first discovery tree and processor mapping.
The first discovery tree creates a tree of clusters, connecting the cluster
from which was originated the trajectory that discovered the child node. The
conformation network is a directed graph that connects all cluster that are
accessible from one another. To do this the processor mapping maps
the propagation trajectories to the cluster they originated.